# openapi/account-service.openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"https://localhost:8081","description":"Generated server url"}],"paths":{"/transfers":{"get":{"tags":["MoneyTransfers"],"operationId":"listMoneyTransfers","parameters":[{"name":"fromIban","in":"query","required":false,"schema":{"type":"string","pattern":"\\d+"}},{"name":"toIban","in":"query","required":false,"schema":{"type":"string","pattern":"\\d+"}},{"name":"minAmount","in":"query","required":false,"schema":{"type":"integer","format":"int64","minimum":0}},{"name":"maxAmount","in":"query","required":false,"schema":{"type":"integer","format":"int64","minimum":0}},{"name":"currencyIso3","in":"query","required":false,"schema":{"type":"string","pattern":"\\w{3}"}},{"name":"timestampAfter","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"timestampBefore","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"labelContaining","in":"query","required":false,"schema":{"type":"string","maxLength":2147483647,"minLength":3}},{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","contains":{"default":""},"items":{"type":"string","default":""},"unevaluatedItems":{"default":""}}}],"responses":{"422":{"description":"Unprocessable Content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationProblemDetail","additionalProperties":{"default":""},"default":""}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"200":{"description":"OK","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/PagedModelMoneyTransferResponse"}}}}}},"post":{"tags":["MoneyTransfers"],"operationId":"transferMoneyBetweenAccounts","parameters":[{"name":"dto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/MoneyTransferRequest"}}],"responses":{"422":{"description":"Unprocessable Content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationProblemDetail","additionalProperties":{"default":""},"default":""}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"202":{"description":"Accepted"}}}},"/accounts":{"get":{"tags":["Accounts"],"operationId":"listAccounts","parameters":[{"name":"customerId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"422":{"description":"Unprocessable Content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationProblemDetail","additionalProperties":{"default":""},"default":""}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"200":{"description":"OK","content":{"application/problem+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"post":{"tags":["Accounts"],"operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCreationRequest"}}},"required":true},"responses":{"422":{"description":"Unprocessable Content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationProblemDetail","additionalProperties":{"default":""},"default":""}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"200":{"description":"OK"}}}},"/transfers/{transferId}":{"get":{"tags":["MoneyTransfers"],"operationId":"getMoneyTransfer","parameters":[{"name":"transferId","in":"path","required":true,"schema":{"type":"integer","default":""}}],"responses":{"422":{"description":"Unprocessable Content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationProblemDetail","additionalProperties":{"default":""},"default":""}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"200":{"description":"OK","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/MoneyTransferResponse"}}}}}}},"/accounts/{iban}":{"get":{"tags":["Accounts"],"operationId":"getAccount","parameters":[{"name":"iban","in":"path","required":true,"schema":{"type":"string","default":""}}],"responses":{"422":{"description":"Unprocessable Content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationProblemDetail","additionalProperties":{"default":""},"default":""}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetail","additionalProperties":{"default":""},"default":""}}}},"200":{"description":"OK","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}}},"components":{"schemas":{"ValidationProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{}}}},"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{}}}},"MoneyTransferRequest":{"type":"object","properties":{"sourceIban":{"type":"string"},"destinationIban":{"type":"string"},"amount":{"type":"integer","format":"int64","minimum":1},"currency":{"type":"string","pattern":"\\w{3}"},"label":{"type":"string","minLength":1}},"required":["destinationIban","label","sourceIban"]},"AccountCreationRequest":{"type":"object","properties":{"iban":{"type":"string"},"customerId":{"type":"integer","format":"int64"},"currency":{"type":"string","pattern":"[A-Z]{3}"}},"required":["currency","customerId","iban"]},"MoneyTransferResponse":{"type":"object","properties":{"fromIban":{"type":"string"},"toIban":{"type":"string"},"amount":{"type":"integer","format":"int64","description":"In minor unit (i.e. 1000 for 1000 XPF, 10.00 USD, 1.000 KWD)"},"currency":{"type":"string","description":"in ISO_3 format"},"timestamp":{"type":"string","format":"date-time"},"label":{"type":"string"}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"}}},"PagedModelMoneyTransferResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/MoneyTransferResponse"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"AccountResponse":{"type":"object","properties":{"iban":{"type":"string"},"customerId":{"type":"integer","format":"int64"},"currency":{"type":"string"},"balance":{"type":"integer","format":"int64"}}}}}}